upd: move pack type selection to seperate tab
