+++ src/components/Projects/CreateProject/Tabs/GeneralTab.vue
@@ -1,5 +1,8 @@
 <template>
 	<div>
+		<!--TODO: make translatable-->
+		<h2 class="mb-2">General</h2>
+
 		<div class="d-flex">
 			<div
 				style="width: 40%"
